Из MainActivityTest вынесена базовая логика в MainActivityTestsBase, создан MainActivitySessionsTest
